-Layout tweaks -Remove several compile warnings
svn path=/branches/extragear/kde3/utils/krecipes/; revision=714517

+++ src/widgets/propertylistview.cpp
@@ -239,7 +239,7 @@ kdDebug() << "saveProp: " << i->text( 0 ) << endl;
 
 bool StdPropertyListView::checkBounds( const QString &name )
 {
-	if ( name.length() > database->maxPropertyNameLength() ) {
+	if ( name.length() > (uint)database->maxPropertyNameLength() ) {
 		KMessageBox::error( this, QString( i18n( "Property name cannot be longer than %1 characters." ) ).arg( database->maxPropertyNameLength() ) );
 		return false;
 	}
